+++ packages/jspsych/src/modules/recording/SessionRecorder.ts
@@ -1,316 +1,42 @@
-import * as randomization from "./randomization";
-
-// ---------------------------------------------------------------------------
-// Schema types (schema_version: 1)
-// ---------------------------------------------------------------------------
-
-export type JsonValue =
-  | null
-  | boolean
-  | number
-  | string
-  | JsonValue[]
-  | { [key: string]: JsonValue };
-
-export interface SessionRecording {
-  schema_version: 1;
-  jspsych_version: string;
-  recording_started_at: string;
-  recording_started_at_perf: number;
-  user_agent: string;
-  viewport: ViewportState;
-  rng: { seed: string | null; math_random_patched: boolean };
-  display_element_id: string;
-  stylesheets: StylesheetSnapshot[];
-  stylesheet_events: StylesheetEvent[];
-  trials: TrialRecording[];
-  viewport_changes: ViewportChange[];
-  // Session-scoped (not per-trial) so RNG calls during pre-trial
-  // parameter evaluation, between-trial gaps, and `on_finish` are
-  // captured. Replayer consumes in order.
-  rng_calls: RngCall[];
-  ended_at_perf: number | null;
-  end_reason: "finished" | "aborted" | "unload" | "memory_limit" | null;
-}
-
-/**
- * Options for `initJsPsych({ record_session })`. Pass `true` for the
- * defaults, `false` to disable, or this object to tune what gets
- * captured. Each `capture_*` flag defaults to `true`; setting one to
- * `false` skips both the listeners and the corresponding event types in
- * the recording.
- */
-export interface RecordSessionOptions {
-  /** Capture form input values (text, textarea, checkbox/radio, select).
-   *  Defaults true. Set false for surveys whose responses must not be
-   *  retained verbatim. */
-  capture_inputs?: boolean;
-  /** Capture canvas pixel snapshots. Defaults true. Set false when
-   *  experiments use large/animated canvases and the pixel readback
-   *  would dominate recording size or perturb timing. */
-  capture_canvas?: boolean;
-  /** Capture every `Math.random()` output to `rng_calls`. Defaults true.
-   *  Set false when reproducibility isn't needed and the RNG is called
-   *  millions of times (e.g. shuffling a large array per trial). */
-  capture_random?: boolean;
-  /** Hard cap on total recorded events across all categories
-   *  (per-trial events + rng_calls + viewport_changes + stylesheet_events).
-   *  When exceeded, recording stops with `end_reason: "memory_limit"`.
-   *  Defaults to no limit. */
-  max_events?: number;
-}
-
-interface ResolvedRecordSessionOptions {
-  capture_inputs: boolean;
-  capture_canvas: boolean;
-  capture_random: boolean;
-  max_events: number;
-}
-
-/**
- * Normalizes the user-facing `record_session` value into either `null`
- * (disabled) or a fully-defaulted options object. Exported for callers
- * (e.g. `JsPsych`) that need to decide whether to construct a recorder.
- */
-export function resolveRecordSessionOptions(
-  raw: boolean | RecordSessionOptions | undefined
-): ResolvedRecordSessionOptions | null {
-  if (!raw) return null;
-  const obj = raw === true ? {} : raw;
-  return {
-    capture_inputs: obj.capture_inputs ?? true,
-    capture_canvas: obj.capture_canvas ?? true,
-    capture_random: obj.capture_random ?? true,
-    max_events: obj.max_events ?? Infinity,
-  };
-}
-
-// A snapshot of one stylesheet attached to the document. `inline` covers
-// `<style>` tags; `link` covers `<link rel="stylesheet">`. `css` is the
-// resolved rule text where readable; cross-origin sheets throw SecurityError
-// on `cssRules` access, so for those we record `href` only and leave `css`
-// null so a replayer can fetch the source out-of-band. `id` is unique within
-// the session and shared with `stylesheet_events` so add/remove/update
-// events can reference snapshots created at start.
-export type StylesheetSnapshot =
-  | { id: number; kind: "inline"; css: string; media: string | null }
-  | { id: number; kind: "link"; href: string; css: string | null; media: string | null };
-
-// Session-scope log entries for `<head>` stylesheet changes after start.
-// `stylesheet.add` carries a full snapshot (with a newly-assigned `id`);
-// `stylesheet.remove` and `stylesheet.update` reference an existing `id`
-// from `stylesheets` or a prior `add` event.
-export type StylesheetEvent =
-  | { type: "stylesheet.add"; t: number; sheet: StylesheetSnapshot }
-  | { type: "stylesheet.remove"; t: number; id: number }
-  | { type: "stylesheet.update"; t: number; id: number; css: string };
-
-export interface ViewportState {
-  w: number;
-  h: number;
-  dpr: number;
-  scale: number;
-  offset_x: number;
-  offset_y: number;
-}
-
-export interface ViewportChange extends ViewportState {
-  t: number;
-}
-
-export interface TrialRecording {
-  trial_index: number;
-  t_start: number;
-  t_dom_ready: number | null;
-  t_end: number | null;
-  plugin: string;
-  initial_dom: DomNode | null;
-  events: RecordedEvent[];
-  trial_data: JsonValue;
-}
-
-export type DomNode = ElementNode | TextNode | CommentNode;
-
-export interface ElementNode {
-  id: number;
-  kind: "element";
-  tag: string;
-  attrs: Record<string, string>;
-  children: DomNode[];
-  canvas_size?: { w: number; h: number };
-  media_src?: string;
-}
-
-export interface TextNode {
-  id: number;
-  kind: "text";
-  text: string;
-}
-
-export interface CommentNode {
-  id: number;
-  kind: "comment";
-  text: string;
-}
-
-export type RecordedEvent =
-  | DomMutation
-  | InputRecord
-  | ClipboardRecord
-  | MediaRecord
-  | FocusRecord
-  | ScrollRecord
-  | CanvasSnapshot;
-
-export type DomMutation =
-  | { type: "dom.add"; t: number; parent: number; before: number | null; node: DomNode }
-  | { type: "dom.remove"; t: number; node: number }
-  | { type: "dom.attr"; t: number; node: number; name: string; value: string | null }
-  | { type: "dom.text"; t: number; node: number; text: string };
-
-export type InputRecord =
-  | { type: "mouse.move"; t: number; x: number; y: number }
-  | {
-      type: "mouse.down" | "mouse.up" | "mouse.click";
-      t: number;
-      x: number;
-      y: number;
-      button: number;
-      target: number | null;
-    }
-  | {
-      type: "touch.start" | "touch.move" | "touch.end";
-      t: number;
-      touches: { id: number; x: number; y: number }[];
-    }
-  | {
-      type: "key.down" | "key.up";
-      t: number;
-      key: string;
-      code: string;
-      mods: { ctrl: boolean; shift: boolean; alt: boolean; meta: boolean };
-      repeat: boolean;
-      target: number | null;
-    }
-  // Form-state changes. The MutationObserver only sees the `value` *attribute*,
-  // not the IDL property the browser writes when a participant types or
-  // toggles a control. These records carry the post-change value/checked/
-  // selection so a replayer can reconstitute survey responses without
-  // having to replay keystrokes through a live form.
-  | { type: "input.value"; t: number; node: number; value: string }
-  | { type: "input.checked"; t: number; node: number; checked: boolean }
-  | { type: "input.select"; t: number; node: number; values: string[] };
-
-export interface ClipboardRecord {
-  type: "clipboard.copy" | "clipboard.cut" | "clipboard.paste";
-  t: number;
-  text: string | null;
-  html: string | null;
-  target: number | null;
-}
-
-export type MediaRecord = {
-  type: "media.play" | "media.pause" | "media.ended" | "media.seeked" | "media.time";
-  t: number;
-  node: number;
-  current_time: number;
-};
-
-export interface FocusRecord {
-  type: "focus" | "blur" | "fullscreen.enter" | "fullscreen.exit";
-  t: number;
-}
-
-export type ScrollRecord =
-  | { type: "scroll.window"; t: number; x: number; y: number }
-  | { type: "scroll.element"; t: number; node: number; x: number; y: number };
-
-// PNG data URL of `<canvas>` pixel state. MutationObserver can't see
-// drawing operations, so the replayer needs these to reconstruct
-// anything participants draw. `region` distinguishes a full-canvas
-// baseline (omitted: drawn at 0,0 after clearing) from an incremental
-// patch (present: composited at region.x,y). First snapshot per canvas
-// is always full; partials only when dirty area < 80% of canvas.
-export interface CanvasSnapshot {
-  type: "canvas.snapshot";
-  t: number;
-  node: number;
-  data_url: string;
-  region?: { x: number; y: number; w: number; h: number };
-}
-
-export interface RngCall {
-  t: number;
-  fn: string;
-  args: JsonValue;
-  result: JsonValue;
-}
-
-// ---------------------------------------------------------------------------
-// SessionRecorder
-// ---------------------------------------------------------------------------
-
-const SCHEMA_VERSION = 1;
-const VIEWPORT_DEBOUNCE_MS = 100;
-const MEDIA_TIME_THROTTLE_MS = 250;
-// Hoisted to avoid allocating a fresh empty array on every Math.random
-// call; the rng_calls log can grow into the millions for a long
-// randomization-heavy session.
-const RNG_NO_ARGS: JsonValue[] = [];
-// Events listened on `<video>` / `<audio>` while a trial is active.
-// Listed once so attach/detach iterate the same set without drift.
-const MEDIA_EVENTS = ["play", "pause", "ended", "seeked", "timeupdate"] as const;
-// Per-canvas minimum gap between gesture-driven snapshots. Bounds the
-// `toDataURL` cost for users who rapidly click/release; trial-end
-// captures bypass this so the final state always lands.
-const CANVAS_SNAPSHOT_MIN_INTERVAL_MS = 250;
-// When the changed bounding box covers more than this fraction of the
-// canvas, fall back to a full snapshot. Cropping a near-full region
-// pays the `getImageData` + `drawImage` + `toDataURL` cost of a full
-// snapshot anyway, so the partial-snapshot bookkeeping is wasted.
-const CANVAS_FULL_SNAPSHOT_AREA_FRACTION = 0.8;
-// Per-canvas minimum gap between draw-triggered animation snapshots.
-// Distinct from `CANVAS_SNAPSHOT_MIN_INTERVAL_MS`, which throttles the
-// gesture-driven path. ~15 Hz is enough fidelity for replaying typical
-// canvas animations without producing 60 PNGs/sec for a continuously-
-// repainting stimulus.
-const CANVAS_ANIMATION_MIN_INTERVAL_MS = 66;
-// Pixel-mutating methods on `CanvasRenderingContext2D`. Path-state
-// methods (`beginPath`, `moveTo`, `lineTo`, …) are intentionally
-// omitted: they don't change pixels until `fill` or `stroke` is called.
-const CANVAS_DRAW_METHODS = [
-  "clearRect",
-  "fillRect",
-  "strokeRect",
-  "fill",
-  "stroke",
-  "fillText",
-  "strokeText",
-  "drawImage",
-  "putImageData",
-] as const;
-
-// Per-canvas snapshot bookkeeping. All fields are optional because they
-// fill in lazily as the canvas first emits, throttles, or animates.
-interface CanvasState {
-  // Last full-canvas data URL. Used to dedupe byte-identical
-  // re-snapshots from the gesture/forced paths.
-  lastSnapshot?: string;
-  // Last time any snapshot was emitted for this canvas. Drives the
-  // gesture-path throttle (`CANVAS_SNAPSHOT_MIN_INTERVAL_MS`).
-  lastSnapshotTime?: number;
-  // Last full-canvas pixel buffer; the next snapshot diffs against
-  // this to find the changed bounding box. Absent for canvases
-  // without a readable 2d context (WebGL); those always emit full
-  // snapshots.
-  shadowData?: ImageData;
-  // Last time the draw-detection animation path emitted, throttled
-  // independently of the gesture path via `CANVAS_ANIMATION_MIN_INTERVAL_MS`.
-  animationLastTime?: number;
-  // Set by the patched 2d-context draw methods; read-and-cleared by
-  // the frame tick. `undefined` and `false` are equivalent.
-  dirty?: boolean;
-}
+import {
+  CANVAS_ANIMATION_MIN_INTERVAL_MS,
+  CANVAS_FULL_SNAPSHOT_AREA_FRACTION,
+  CANVAS_SNAPSHOT_MIN_INTERVAL_MS,
+  CanvasState,
+  MEDIA_EVENTS,
+  MEDIA_TIME_THROTTLE_MS,
+  RNG_NO_ARGS,
+  SCHEMA_VERSION,
+  VIEWPORT_DEBOUNCE_MS,
+} from "./constants";
+import {
+  getReadable2dContext,
+  registerCanvasRecorder,
+  registerMathRandomRecorder,
+  unregisterCanvasRecorder,
+  unregisterMathRandomRecorder,
+} from "./global-patches";
+import {
+  computeDiffBbox,
+  cropCanvasToDataURL,
+  readMedia,
+  readSheetText,
+  readViewport,
+  serializeJson,
+} from "./helpers";
+import type { ResolvedRecordSessionOptions } from "./options";
+import type {
+  ClipboardRecord,
+  DomNode,
+  ElementNode,
+  FocusRecord,
+  MediaRecord,
+  RecordedEvent,
+  SessionRecording,
+  StylesheetSnapshot,
+  TrialRecording,
+  ViewportState,
+} from "./types";
 
 export interface SessionRecorderOptions {
   jspsychVersion: string;
@@ -377,8 +103,7 @@ export class SessionRecorder {
 
   // Per-canvas state: strong-ref Map (we explicitly delete entries on
   // canvas removal and trial end) so iteration and lookup are one
-  // structure. Replaces five parallel WeakMaps that all keyed off the
-  // same canvas element.
+  // structure.
   private canvasStates = new Map<HTMLCanvasElement, CanvasState>();
   // True when an in-flight RAF-deferred sweep should bypass the per-
   // canvas throttle. Set by initial-baseline schedules; gesture-driven
@@ -690,9 +415,8 @@ export class SessionRecorder {
       this.mutationObserver = null;
     }
     this.detachMediaListeners();
-    // WeakMap entries for the released canvases will be GC'd once the
-    // Strong refs drop here; explicit deletion happens in
-    // `releaseSubtree` for canvases removed mid-trial.
+    // Strong refs drop here; explicit deletion happens in `releaseSubtree`
+    // for canvases removed mid-trial.
     this.canvasStates.clear();
     this.pendingCanvasSnapshotForce = false;
     // Tear down only the trial-scoped handlers; session-scoped handlers
@@ -1513,329 +1237,4 @@ export class SessionRecorder {
   private t(): number {
     return performance.now() - this.startPerf;
   }
-}
-
-// ---------------------------------------------------------------------------
-// Module-level patch coordination
-// ---------------------------------------------------------------------------
-// Multiple `JsPsych` instances can run with `record_session: true`
-// concurrently. The previous implementation patched
-// `CanvasRenderingContext2D.prototype` and `Math.random` per-recorder,
-// which meant the second recorder's "original" was actually the first
-// recorder's wrapper; restoring on stop in interleaved order left the
-// process with stale wrappers reinstalled on the prototype. The
-// helpers below install each global patch exactly once (refcounted by
-// the set of active recorders) so registration/unregistration order
-// is irrelevant.
-
-const activeCanvasRecorders: Set<SessionRecorder> = new Set();
-let canvasMethodOriginals: Map<string, Function> | null = null;
-// Per-canvas committed context type, learned by wrapping
-// `HTMLCanvasElement.prototype.getContext`. The recorder consults
-// this map to decide whether `getContext("2d")` is safe to call —
-// see `getReadable2dContext` for the rationale.
-const committedContextTypes: WeakMap<HTMLCanvasElement, string> = new WeakMap();
-let getContextOriginal: HTMLCanvasElement["getContext"] | null = null;
-
-function registerCanvasRecorder(recorder: SessionRecorder) {
-  if (activeCanvasRecorders.has(recorder)) return;
-  installCanvasGlobalPatches();
-  activeCanvasRecorders.add(recorder);
-}
-
-function unregisterCanvasRecorder(recorder: SessionRecorder) {
-  if (!activeCanvasRecorders.delete(recorder)) return;
-  if (activeCanvasRecorders.size === 0) uninstallCanvasGlobalPatches();
-}
-
-function installCanvasGlobalPatches() {
-  if (canvasMethodOriginals !== null) return;
-  if (typeof CanvasRenderingContext2D !== "undefined") {
-    const proto = CanvasRenderingContext2D.prototype as unknown as Record<string, Function>;
-    canvasMethodOriginals = new Map();
-    for (const method of CANVAS_DRAW_METHODS) {
-      const original = proto[method];
-      if (typeof original !== "function") continue;
-      canvasMethodOriginals.set(method, original);
-      proto[method] = function (this: CanvasRenderingContext2D, ...args: unknown[]) {
-        try {
-          const canvas = this.canvas as HTMLCanvasElement | undefined;
-          if (canvas) {
-            for (const r of activeCanvasRecorders) r.notifyCanvasDraw(canvas);
-          }
-        } catch {
-          // Tracking must never break the experiment's own drawing.
-        }
-        return (original as Function).apply(this, args);
-      };
-    }
-  }
-  // Wrap `getContext` so we learn each canvas's committed context type
-  // without ever creating a context ourselves. This is what lets
-  // `getReadable2dContext` skip the diff path for WebGL canvases — and,
-  // crucially, for canvases that the user hasn't yet committed to any
-  // type, since calling `getContext("2d")` on a fresh canvas locks it
-  // into 2D and breaks any later `getContext("webgl")` call.
-  if (typeof HTMLCanvasElement !== "undefined") {
-    const proto = HTMLCanvasElement.prototype;
-    getContextOriginal = proto.getContext;
-    const wrapped = function (this: HTMLCanvasElement, ...args: unknown[]) {
-      const result = (getContextOriginal as Function).apply(this, args);
-      if (result && !committedContextTypes.has(this) && typeof args[0] === "string") {
-        committedContextTypes.set(this, args[0] as string);
-      }
-      return result;
-    };
-    proto.getContext = wrapped as typeof proto.getContext;
-  }
-}
-
-function uninstallCanvasGlobalPatches() {
-  if (canvasMethodOriginals !== null && typeof CanvasRenderingContext2D !== "undefined") {
-    const proto = CanvasRenderingContext2D.prototype as unknown as Record<string, Function>;
-    for (const [method, original] of canvasMethodOriginals) {
-      proto[method] = original;
-    }
-  }
-  canvasMethodOriginals = null;
-  if (getContextOriginal !== null && typeof HTMLCanvasElement !== "undefined") {
-    HTMLCanvasElement.prototype.getContext = getContextOriginal;
-  }
-  getContextOriginal = null;
-}
-
-const activeMathRandomRecorders: Set<SessionRecorder> = new Set();
-let mathRandomPrePatch: typeof Math.random | null = null;
-let mathRandomWrapper: typeof Math.random | null = null;
-
-function isNativeMathRandom(fn: typeof Math.random) {
-  // Native built-ins stringify with the [native code] sentinel; any
-  // user-installed PRNG (e.g. via jsPsych.randomization.setSeed) will
-  // serialize as JS source instead.
-  return /\{\s*\[native code\]\s*\}/.test(Function.prototype.toString.call(fn));
-}
-
-// Returns the seed string assigned by auto-seeding when this recorder
-// caused the install, or null when no seeding happened (either because
-// the user already installed a PRNG, or because a previous recorder
-// already patched).
-function registerMathRandomRecorder(recorder: SessionRecorder): string | null {
-  if (activeMathRandomRecorders.has(recorder)) return null;
-  let seed: string | null = null;
-  if (mathRandomWrapper === null) {
-    mathRandomPrePatch = Math.random;
-    if (isNativeMathRandom(mathRandomPrePatch)) {
-      seed = randomization.setSeed();
-    }
-    const upstream = Math.random.bind(Math);
-    mathRandomWrapper = () => {
-      const v = upstream();
-      // Captured at session scope on every active recorder so calls
-      // outside trial boundaries (pre-trial parameter evaluation,
-      // post-trial gaps, the experimenter's `on_finish`) are still
-      // recorded.
-      for (const r of activeMathRandomRecorders) r.recordRngCall(v);
-      return v;
-    };
-    Math.random = mathRandomWrapper;
-  }
-  activeMathRandomRecorders.add(recorder);
-  return seed;
-}
-
-function unregisterMathRandomRecorder(recorder: SessionRecorder) {
-  if (!activeMathRandomRecorders.delete(recorder)) return;
-  if (activeMathRandomRecorders.size > 0) return;
-  // Only restore if our wrapper is still installed. If the user
-  // reassigned `Math.random` mid-session, leave their replacement in
-  // place — clobbering it would silently undo a deliberate user
-  // decision.
-  if (mathRandomWrapper !== null && Math.random === mathRandomWrapper) {
-    if (mathRandomPrePatch !== null) Math.random = mathRandomPrePatch;
-  }
-  mathRandomWrapper = null;
-  mathRandomPrePatch = null;
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-// Reads the `media` attribute, preferring the parsed value on the
-// CSSStyleSheet (which normalizes whitespace/casing) and falling back to
-// the raw attribute on the owning element.
-function readMedia(el: HTMLElement, sheet: CSSStyleSheet | null): string | null {
-  const fromSheet = sheet?.media?.mediaText;
-  if (fromSheet) return fromSheet;
-  const attr = el.getAttribute("media");
-  return attr && attr.length > 0 ? attr : null;
-}
-
-// Reads the resolved CSS rule text from a stylesheet. Returns null when
-// `cssRules` is unreadable (cross-origin sheets without CORS access throw
-// SecurityError) so callers can record only the href in that case.
-function readSheetText(sheet: CSSStyleSheet): string | null {
-  try {
-    const rules = sheet.cssRules;
-    if (!rules) return null;
-    const parts: string[] = [];
-    for (let i = 0; i < rules.length; i++) {
-      parts.push(rules[i].cssText);
-    }
-    return parts.join("\n");
-  } catch {
-    return null;
-  }
-}
-
-// Returns the canvas's 2d rendering context if one was already
-// committed by user code, otherwise null. Calling `getContext("2d")`
-// on a fresh canvas would permanently lock it to 2D and break any
-// subsequent `getContext("webgl")` call from a plugin, so we only
-// touch canvases whose committed type was learned via the
-// `HTMLCanvasElement.prototype.getContext` wrapper installed in
-// `installCanvasGlobalPatches`. Canvases of unknown or non-2d type
-// fall through to full `toDataURL` snapshots in `snapshotCanvas`.
-function getReadable2dContext(canvas: HTMLCanvasElement): CanvasRenderingContext2D | null {
-  if (committedContextTypes.get(canvas) !== "2d") return null;
-  try {
-    const ctx = canvas.getContext("2d");
-    return ctx ?? null;
-  } catch {
-    return null;
-  }
-}
-
-// Finds the bounding box of pixels that differ between two same-size
-// ImageData buffers. Uses an edge-shrink walk: scan top-down for the
-// first dirty row, bottom-up for the last, then within that row band
-// scan inward from each side for the first dirty column. For typical
-// incremental drawing (a stroke, a localized clear) this terminates
-// after only a few thousand pixel comparisons even on a multi-megapixel
-// canvas. Returns null when the buffers are byte-identical.
-//
-// Exported for unit testing; a replayer doesn't need it.
-export function computeDiffBbox(
-  prev: Uint8ClampedArray,
-  curr: Uint8ClampedArray,
-  w: number,
-  h: number
-): { x: number; y: number; w: number; h: number } | null {
-  let top = -1;
-  for (let y = 0; y < h; y++) {
-    const rowStart = y * w * 4;
-    const rowEnd = rowStart + w * 4;
-    for (let i = rowStart; i < rowEnd; i++) {
-      if (prev[i] !== curr[i]) {
-        top = y;
-        break;
-      }
-    }
-    if (top !== -1) break;
-  }
-  if (top === -1) return null;
-
-  let bottom = top;
-  for (let y = h - 1; y > top; y--) {
-    const rowStart = y * w * 4;
-    const rowEnd = rowStart + w * 4;
-    let dirty = false;
-    for (let i = rowStart; i < rowEnd; i++) {
-      if (prev[i] !== curr[i]) {
-        dirty = true;
-        break;
-      }
-    }
-    if (dirty) {
-      bottom = y;
-      break;
-    }
-  }
-
-  let left = w - 1;
-  let right = 0;
-  for (let y = top; y <= bottom; y++) {
-    const rowStart = y * w * 4;
-    // Scan inward from the left up to the current `left` candidate.
-    for (let x = 0; x < left; x++) {
-      const i = rowStart + x * 4;
-      if (
-        prev[i] !== curr[i] ||
-        prev[i + 1] !== curr[i + 1] ||
-        prev[i + 2] !== curr[i + 2] ||
-        prev[i + 3] !== curr[i + 3]
-      ) {
-        left = x;
-        break;
-      }
-    }
-    // Scan inward from the right past the current `right` candidate.
-    for (let x = w - 1; x > right; x--) {
-      const i = rowStart + x * 4;
-      if (
-        prev[i] !== curr[i] ||
-        prev[i + 1] !== curr[i + 1] ||
-        prev[i + 2] !== curr[i + 2] ||
-        prev[i + 3] !== curr[i + 3]
-      ) {
-        right = x;
-        break;
-      }
-    }
-    if (left === 0 && right === w - 1) break;
-  }
-
-  return { x: left, y: top, w: right - left + 1, h: bottom - top + 1 };
-}
-
-// Copies a rectangular region out of a source canvas into a temporary
-// canvas and returns its data URL. The temporary canvas is sized to
-// the region so the resulting PNG carries only the dirty pixels.
-function cropCanvasToDataURL(
-  source: HTMLCanvasElement,
-  region: { x: number; y: number; w: number; h: number }
-): string {
-  const tmp = document.createElement("canvas");
-  tmp.width = region.w;
-  tmp.height = region.h;
-  const tctx = tmp.getContext("2d");
-  if (!tctx) return source.toDataURL();
-  tctx.drawImage(source, region.x, region.y, region.w, region.h, 0, 0, region.w, region.h);
-  return tmp.toDataURL();
-}
-
-function readViewport(): ViewportState {
-  const vv = window.visualViewport;
-  return {
-    w: window.innerWidth,
-    h: window.innerHeight,
-    dpr: window.devicePixelRatio || 1,
-    scale: vv?.scale ?? 1,
-    offset_x: vv?.offsetLeft ?? 0,
-    offset_y: vv?.offsetTop ?? 0,
-  };
-}
-
-/**
- * JSON-safe serialization for `trial_data` (the per-trial data row). Drops
- * functions and DOM nodes — the recording's visual fidelity comes from DOM
- * mutations, not from re-executing trial parameters.
- */
-export function serializeJson(value: unknown, seen = new WeakSet<object>()): JsonValue {
-  if (value === null || value === undefined) return null;
-  const t = typeof value;
-  if (t === "boolean" || t === "string") return value as JsonValue;
-  if (t === "number") return Number.isFinite(value as number) ? (value as number) : null;
-  if (t === "function") return null;
-  if (t !== "object") return null;
-  if (seen.has(value as object)) return null;
-  seen.add(value as object);
-  if (Array.isArray(value)) return value.map((v) => serializeJson(v, seen));
-  if (value instanceof Date) return value.toISOString();
-  if (value instanceof Element || value instanceof Node) return null;
-  const out: Record<string, JsonValue> = {};
-  for (const [k, v] of Object.entries(value as Record<string, unknown>)) {
-    out[k] = serializeJson(v, seen);
-  }
-  return out;
 }
